app/requirements.txt
Django
django-rest-framework
gunicorn